move all files from modules directory to grading directory & bugfix on svn checkout
