+++ grading/STRutil.h
