+++ grading/difference.cpp
