+++ grading/modules.h
@@ -1,0 +1,28 @@
+/* FILENAME: modules.h
+ * YEAR: 2014
+ * AUTHORS: Please refer to 'AUTHORS.md' for a list of contributors
+ * LICENSE: Please refer to 'LICENSE.md' for the conditions of using this code
+ *
+ * RELEVANT DOCUMENTATION:
+ */
+
+#ifndef __MODULES__
+#define __MODULES__
+//#include "diffNaive.h"  /* this is now a .cpp file */
+
+//#include "diffNaive.cpp"  /* this is now a .cpp file */
+
+#include "STRutil.h"
+#include "tokenSearch.h"
+
+//#include "tokenSearch.cpp"  /* should not #include a .cpp file */
+
+#include "myersDiff.h"
+#include "textMasking.h"
+#include "clean.h"
+ 
+#include "difference.h"
+#include "testResults.h"
+#include "tokens.h"
+
+#endif //__MODULES__
