+++ grading/runDiff.h
@@ -24,7 +24,7 @@
 #include <string>
 #include <vector>
 #include <fstream>
-#include "modules/modules.h"
+#include "modules.h"
 
 /* METHOD: readFileList
  * ARGS: input: list of file names, sample_file: instructor file,
