+++ grading/textMasking.cpp
