--- modules/modules.h
@@ -1,28 +1,0 @@
-/* FILENAME: modules.h
- * YEAR: 2014
- * AUTHORS: Please refer to 'AUTHORS.md' for a list of contributors
- * LICENSE: Please refer to 'LICENSE.md' for the conditions of using this code
- *
- * RELEVANT DOCUMENTATION:
- */
-
-#ifndef __MODULES__
-#define __MODULES__
-//#include "modules/diffNaive.h"  /* this is now a .cpp file */
-
-//#include "modules/diffNaive.cpp"  /* this is now a .cpp file */
-
-#include "modules/STRutil.h"
-#include "modules/tokenSearch.h"
-
-//#include "modules/tokenSearch.cpp"  /* should not #include a .cpp file */
-
-#include "modules/myersDiff.h"
-#include "modules/textMasking.h"
-#include "modules/clean.h"
- 
-#include "modules/difference.h"
-#include "modules/testResults.h"
-#include "modules/tokens.h"
-
-#endif //__MODULES__
